Remove Mbx prefix
