+++ src/components/button/index.js
@@ -1,0 +1,3 @@
+import main from './button';
+
+export default main;
