+++ src/components/icon-text/examples/icon-text-b.js
@@ -1,0 +1,15 @@
+/*
+Go see the next thing!
+*/
+import React from 'react';
+import IconText from '../icon-text';
+
+export default class Example extends React.Component {
+  render() {
+    return (
+      <button type="button" className="link link-purple">
+        <IconText iconAfter="chevron-right">Take me there</IconText>
+      </button>
+    );
+  }
+}
