+++ src/components/icon-text/index.js
@@ -1,0 +1,3 @@
+import main from './icon-text';
+
+export default main;
