+++ src/components/underline-tabs/index.js
@@ -1,0 +1,3 @@
+import main from './underline-tabs';
+
+export default main;
